requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.13
langgraph>=0.0.20
openai>=1.12.0
pymongo>=4.6.1
pinecone-client>=3.0.2
python-dotenv>=1.0.0
chromadb>=0.4.22
unstructured>=0.12.3 